requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.66
botocore==1.42.66
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
Django==6.0.3
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.11
jmespath==1.1.0
kombu==5.6.2
oauthlib==3.3.1
packaging==26.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python3-openid==3.2.0
redis==7.3.0
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.16.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
